# backend/docs/swagger.yaml
basePath: /
definitions:
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handler.authResponse:
    properties:
      token:
        type: string
      user: {}
    type: object
...
  handler.googleLoginRequest:
    properties:
      idToken:
        type: string
    type: object
  handler.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handler.registerRequest:
    properties:
      dormBuilding:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  handler.startConversationRequest:
    properties:
      listingId:
        type: string
    type: object
...
  models.Conversation:
    properties:
      buyerId:
        type: string
      createdAt:
        type: string
      id:
        type: string
      lastMessage:
        $ref: '#/definitions/models.Message'
      lastMessageAt:
        type: string
      listing:
        $ref: '#/definitions/models.Listing'
      listingId:
        type: string
      otherParty:
        allOf:
        - $ref: '#/definitions/models.PublicUser'
        description: เติมตอน list ให้ frontend ไม่ต้องเดาเองว่าใครคือ "อีกฝ่าย"
      sellerId:
        type: string
    type: object
  models.Listing:
    properties:
      categoryId:
        type: string
      condition:
        $ref: '#/definitions/models.ListingCondition'
      createdAt:
        type: string
      deletedAt:
        description: nil = ยังไม่ถูกลบ
        type: string
      description:
        type: string
      id:
        type: string
      images:
        items:
          $ref: '#/definitions/models.ListingImage'
        type: array
      price:
        description: หน่วยเป็นบาท เก็บเป็นจำนวนเต็มเพื่อเลี่ยงปัญหา float
        type: integer
      seller:
        $ref: '#/definitions/models.PublicUser'
      sellerId:
        type: string
      status:
        $ref: '#/definitions/models.ListingStatus'
      suggestedPrice:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
    type: object
  models.ListingCondition:
    enum:
    - new
    - like_new
    - good
    - worn
    type: string
    x-enum-varnames:
    - ConditionNew
    - ConditionLikeNew
    - ConditionGood
    - ConditionWorn
  models.ListingImage:
    properties:
      id:
        type: string
      listingId:
        type: string
      sortOrder:
        type: integer
      url:
        type: string
    type: object
  models.ListingStatus:
    enum:
    - available
    - reserved
    - sold
    type: string
    x-enum-varnames:
    - StatusAvailable
    - StatusReserved
    - StatusSold
  models.Message:
    properties:
      content:
        type: string
      conversationId:
        type: string
      createdAt:
        type: string
      id:
        type: string
      senderId:
        type: string
    type: object
  models.PublicUser:
    properties:
      avatarUrl:
        type: string
      dormBuilding:
        type: string
      id:
        type: string
      name:
        type: string
      trustScore:
        type: integer
    type: object
...
  models.User:
    properties:
      avatarUrl:
        type: string
      createdAt:
        type: string
      dormBuilding:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      trustScore:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: tnswnb@gmail.com
    name: DormMarket Support
  description: REST API สำหรับแพลตฟอร์มตลาดซื้อขายของมือสองในหอ/มหาลัย (DormMarket)
  license:
    name: MIT
  title: DormMarket API
  version: "1.0"
paths:
  /api/auth/google:
    post:
      consumes:
      - application/json
      description: รับ ID token จาก Google Identity Services (frontend ใช้ Google
        Sign-In button ได้ token นี้มาโดยตรง ไม่ต้อง redirect ไปมา) แลกเป็น JWT token
        ของระบบ
      parameters:
      - description: Google ID token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.googleLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.authResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "501":
          description: Google Login ยังไม่เปิดใช้งาน
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: เข้าสู่ระบบด้วย Google
      tags:
      - auth
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: เข้าสู่ระบบด้วยอีเมล/รหัสผ่าน คืน JWT token
      parameters:
      - description: อีเมลและรหัสผ่าน
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.authResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: เข้าสู่ระบบ
      tags:
      - auth
  /api/auth/me:
    get:
      description: คืนข้อมูลผู้ใช้ที่ login อยู่ ตาม JWT token ที่แนบมา
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ข้อมูลผู้ใช้ปัจจุบัน
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: สร้างบัญชีผู้ใช้ด้วยอีเมล/รหัสผ่าน แล้วคืน JWT token กลับมาทันที
      parameters:
      - description: ข้อมูลสมัครสมาชิก
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.authResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: สมัครสมาชิกใหม่
      tags:
      - auth
...
  /api/conversations:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Conversation'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: รายการบทสนทนาของฉัน (inbox)
      tags:
      - chat
    post:
      consumes:
      - application/json
      description: เรียกตอนกดปุ่ม "ติดต่อผู้ขาย" — idempotent เรียกซ้ำกี่ครั้งก็ได้ห้องเดิม
      parameters:
      - description: รหัสประกาศที่จะเริ่มคุย
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.startConversationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Conversation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: เริ่มบทสนทนากับผู้ขาย
      tags:
      - chat
  /api/conversations/{id}:
    get:
      description: คืนข้อมูล conversation พร้อม listing แนบมาด้วย ใช้ตอนเปิดหน้าแชท
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Conversation'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: รายละเอียดบทสนทนา
      tags:
      - chat
  /api/conversations/{id}/messages:
    get:
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Message'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: รายการข้อความในบทสนทนา
      tags:
      - chat
...
securityDefinitions:
  BearerAuth:
    description: พิมพ์ "Bearer" ตามด้วยเว้นวรรคแล้วตามด้วย JWT token เช่น "Bearer
      eyJhbGc..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"